WEAT-5 Organize packaging
