+++ app/src/main/java/com/app/weather/presentation/util/Extensions.kt
@@ -3,3 +3,5 @@ package com.app.weather.presentation.weather
 import android.location.Location
 
 fun Location.getLatLong() = "${this.latitude},${this.longitude}"
+
+fun String.getURL() = "https://$this"
