+++ core/src/main/java/com/app/core/data/network/api/weather/WeatherModule.kt
@@ -7,7 +7,6 @@ import dagger.hilt.components.SingletonComponent
 import retrofit2.Retrofit
 import javax.inject.Singleton
 
-
 @Module
 @InstallIn(SingletonComponent::class)
 object WeatherModule {
