+++ core/src/main/java/com/app/core/data/repository/BaseDataSource.kt
@@ -1,6 +1,5 @@
 package com.app.core.data.repository
 
-import android.util.Log
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOn
@@ -9,15 +8,10 @@ import retrofit2.Response
 open class BaseDataSource {
 
     suspend fun <T : Any> flowOnIO(api: suspend () -> Response<T>) =
-        flow<ResultWrapper<T>> {
+        flow {
             try {
                 val response = api.invoke()
-//                val body = handleResponse(response)
-
                 if (response.isSuccessful) {
-
-//                    response.body()
-
                     response.body()?.let {
                         emit(ResultWrapper.Success(it))
                     }
@@ -26,24 +20,10 @@ open class BaseDataSource {
                         emit(ResultWrapper.ErrorString(response.message()))
                     }
                 }
-
-//                body?.let { emit(it) }
             } catch (e: Exception) {
-                Log.i("flow_tag", "flowOnIO: ${e.message}")
                 ResultWrapper.Success("Unknown Error occurred!")
             }
         }.flowOn(Dispatchers.IO)
-
-
-//    private fun <T> handleResponse(response: Response<T>): Result {
-//        if (response.code() == 200) {
-//
-//            return ResultWrapper.Success(response.message())
-//        }
-//        return response.body()
-//    }
-
-
 }
