+++ core/src/main/java/com/app/core/data/repository/home/WeatherRepositoryImpl.kt
@@ -6,8 +6,9 @@ import com.app.core.domain.ForecastResponse
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
-
-class WeatherRepositoryImpl @Inject constructor(private val weatherDataSource: WeatherDataSource) : WeatherRepository {
+class WeatherRepositoryImpl @Inject constructor(
+    private val weatherDataSource: WeatherDataSource
+) : WeatherRepository {
     override suspend fun forecast(latLong: String): Flow<ResultWrapper<ForecastResponse>> {
         return weatherDataSource.forecast(latLong)
     }
