+++ core/src/main/java/com/app/core/interactor/weather/ForecastInteractor.kt
@@ -3,7 +3,6 @@ package com.app.core.interactor.weather
 import com.app.core.data.repository.home.WeatherRepository
 import javax.inject.Inject
 
-
 class ForecastInteractor @Inject constructor(private val repository: WeatherRepository) {
     suspend fun forecast(latLong: String) = repository.forecast(latLong)
 }
