+++ core/src/main/java/com/app/core/interactor/weather/WeatherInteractors.kt
@@ -2,6 +2,5 @@ package com.app.core.interactor.weather
 
 import javax.inject.Inject
 
-
 data class WeatherInteractors @Inject constructor(val forecastInteractor: ForecastInteractor)
